git-svn-id: https://thelia.svn.sourceforge.net/svnroot/thelia@562 6d4e0a53-8826-0410-8364-8b493f4a4454

+++ thelia/classes/Produit.class.php
@@ -99,6 +99,8 @@
 				
 		function supprimer(){
 
+			if ($this->id == 0 || $this->id == "") return;
+
 			$stock = new Stock();
 			$query = "delete from $stock->table where produit='" . $this->id . "'"; 
 			$resul = mysql_query($query, $stock->link);
